requirements.txt
python-dotenv
langchain
langchain-google
langchain-openai
langchain-community
langchain-chroma
chromadb
openai
google-generativeai
langchain-google-genai
pypdf
Flask
Flask-Cors
python-dotenv 
langchain-chroma 
langchain-google-genai 
langchain-core 
pypdf